Tested all the endpoints and fixed small bugs found. API is now working authorization, updating, adding, deleting, etc...

+++ routes/comments.js
@@ -1,19 +1,21 @@
-const { formatSuccess, formatError } = require('../utils/respFormat');
-const router = require('express').Router();
-const db = require('../utils/database');
-const _ = require('lodash');
+const { formatSuccess, formatError } = require("../utils/respFormat");
+const router = require("express").Router();
+const db = require("../utils/database");
+const _ = require("lodash");
 
 // sample req
 // localhost:3000/comment?id=1
 async function getComment(req, res) {
-    let commentID = req.query.id;
-    const comment = await db.getComment(commentID);
+    const postID = req.query.id;
+    const comments = await db.getPostComments(postID);
 
-    if (!comment)
-        return res.status(404).json(formatError(`No comment found with id: ${commentID}`));
+    if (!comments)
+        return res
+            .status(404)
+            .json(formatError(`No comments found for post: ${postID}`));
 
-    res.status(200).json(formatSuccess(comment));
-};
+    res.status(200).json(formatSuccess(comments));
+}
 
 // sample json
 // {
@@ -22,10 +24,10 @@ async function getComment(req, res) {
 //     "post_id": 2
 // }
 async function addComment(req, res) {
-    const comment = _.pick(req.body, ['content', 'post_id']);
+    const comment = _.pick(req.body, ["content", "post_id"]);
     await db.addComment(comment.content, req.user.user_id, comment.post_id);
     res.status(201).json(formatSuccess("Comment added."));
-};
+}
 
 // sample json
 // {
@@ -33,22 +35,38 @@ async function addComment(req, res) {
 //     "comment_id": 1
 // }
 async function updateComment(req, res) {
-    const comment = _.pick(req.body, ['content', 'comment_id']);
-    await db.updateComment(content, comment_id);
+    const newComment = _.pick(req.body, ["content", "comment_id"]);
+    const comment = await db.getComment(newComment.comment_id);
+
+    if (!comment)
+        return res.status(404).json(formatError(`No comment found with id: ${newComment.comment_id}`));
+
+    if (!req.user.is_admin && req.user.user_id != comment.user_id)
+        return res.status(401).json(formatError("This is not your comment."));
+
+    await db.updateComment(newComment.content, comment.comment_id);
     res.status(200).json(formatSuccess("Comment updated."));
-};
+}
 
 // sample req
 // localhost:3000/comment?id=1
 async function deleteComment(req, res) {
-    let comment_id = req.query.id;
-    await db.deleteComment(comment_id);
-    res.status(200).json(formatSuccess("Comment deleted."))
-};
+    const comment_id = req.query.id;
+    const comment = await db.getComment(comment_id);
 
-router.get('/', getComment);
-router.post('/', addComment);
-router.put('/', updateComment);
-router.delete('/', deleteComment);
+    if (!comment)
+        return res.status(404).json(formatError(`No comment found with id: ${comment_id}`));
+
+    if (!req.user.is_admin && req.user.user_id != comment.user_id)
+        return res.status(401).json(formatError("This is not your comment."));
+
+    await db.deleteComment(comment_id);
+    res.status(200).json(formatSuccess("Comment deleted."));
+}
+
+router.get("/", getComment);
+router.post("/", addComment);
+router.put("/", updateComment);
+router.delete("/", deleteComment);
 
 module.exports = router;
