+++ routes/posts.js
@@ -11,9 +11,9 @@ async function getPost(req, resp) {
     let post = await db.getPost(post_id);
 
     if (!post)
-        return res.status(404).json(formatError(`No post found with id: ${post_id}`));
+        return resp.status(404).json(formatError(`No post found with id: ${post_id}`));
 
-    res.status(200).json(formatSuccess(post));
+    resp.status(200).json(formatSuccess(post));
 };
 
 async function getAllPosts(req, resp) {
@@ -31,8 +31,8 @@ async function getAllPosts(req, resp) {
 
 // }
 function addPost(req, resp) {
-    let post = _.pick(req.body, ['post_date', 'title', 'topic', 'content', 'user_id']);
-    db.createPost(post.post_date, post.title, post.topic, post.content, post.user_id);
+    let post = _.pick(req.body, ['post_date', 'title', 'topic', 'content']);
+    db.createPost(post.post_date, post.title, post.topic, post.content, req.user.user_id);
     resp.status(200).json(formatSuccess("Post created."))
 };
 
@@ -43,16 +43,26 @@ function addPost(req, resp) {
 //     "content": "updated content",
 //     "post_id": 2
 // }
-function updatePost(req, resp) {
-    let post = _.pick(req.body, ['title', 'topic', 'content', 'post_id']);
-    db.updatePost(post.title, post.topic, post.content, post.post_id);
+async function updatePost(req, resp) {
+    const newPost = _.pick(req.body, ['title', 'topic', 'content', 'post_id']);
+    const originalPost = await db.getPost(newPost.post_id);
+
+    if (!req.user.is_admin && req.user.user_id != originalPost.user_id)
+        return resp.status(401).json(formatError("This is not your post."));
+
+    db.updatePost(newPost.title, newPost.topic, newPost.content, newPost.post_id);
     resp.status(200).json(formatSuccess('Post updated.'))
 };
 
 // sample req
 // localhost:3000/post?id=1
-function deletePost(req, resp) {
+async function deletePost(req, resp) {
     let post_id = req.query.id;
+    const originalPost = await db.getPost(newPost.post_id);
+
+    if (!req.user.is_admin && req.user.user_id != originalPost.user_id)
+        return resp.status(401).json(formatError("This is not your post."));
+
     db.deletePost(post_id);
     resp.status(200).json(formatSuccess('Post deleted.'))
 };
